+++ xrpl/models/__init__.py
@@ -1,6 +1,7 @@
 """Top-level exports for the models package."""
 from xrpl.models import amounts, currencies, requests, transactions
 from xrpl.models.amounts import *  # noqa: F401, F403
+from xrpl.models.auth_account import AuthAccount
 from xrpl.models.currencies import *  # noqa: F401, F403
 from xrpl.models.exceptions import XRPLModelException
 from xrpl.models.path import Path, PathStep
@@ -8,11 +9,13 @@ from xrpl.models.requests import *  # noqa: F401, F403
 from xrpl.models.response import Response
 from xrpl.models.transactions import *  # noqa: F401, F403
 from xrpl.models.transactions.pseudo_transactions import *  # noqa: F401, F403
+from xrpl.models.xchain_bridge import XChainBridge
 
 __all__ = [
     "XRPLModelException",
     "amounts",
     *amounts.__all__,
+    "AuthAccount",
     "currencies",
     *currencies.__all__,
     "requests",
@@ -23,4 +26,5 @@ __all__ = [
     "Path",
     "PathStep",
     "Response",
+    "XChainBridge",
 ]