+++ xrpl/models/transactions/__init__.py
@@ -9,6 +9,19 @@ from xrpl.models.transactions.account_set import (
     AccountSetFlag,
     AccountSetFlagInterface,
 )
+from xrpl.models.transactions.amm_bid import AMMBid, AuthAccount
+from xrpl.models.transactions.amm_create import AMMCreate
+from xrpl.models.transactions.amm_deposit import (
+    AMMDeposit,
+    AMMDepositFlag,
+    AMMDepositFlagInterface,
+)
+from xrpl.models.transactions.amm_vote import AMMVote
+from xrpl.models.transactions.amm_withdraw import (
+    AMMWithdraw,
+    AMMWithdrawFlag,
+    AMMWithdrawFlagInterface,
+)
 from xrpl.models.transactions.check_cancel import CheckCancel
 from xrpl.models.transactions.check_cash import CheckCash
 from xrpl.models.transactions.check_create import CheckCreate
@@ -53,12 +66,40 @@ from xrpl.models.transactions.trust_set import (
     TrustSetFlag,
     TrustSetFlagInterface,
 )
+from xrpl.models.transactions.xchain_account_create_commit import (
+    XChainAccountCreateCommit,
+)
+from xrpl.models.transactions.xchain_add_account_create_attestation import (
+    XChainAddAccountCreateAttestation,
+)
+from xrpl.models.transactions.xchain_add_claim_attestation import (
+    XChainAddClaimAttestation,
+)
+from xrpl.models.transactions.xchain_claim import XChainClaim
+from xrpl.models.transactions.xchain_commit import XChainCommit
+from xrpl.models.transactions.xchain_create_bridge import XChainCreateBridge
+from xrpl.models.transactions.xchain_create_claim_id import XChainCreateClaimID
+from xrpl.models.transactions.xchain_modify_bridge import (
+    XChainModifyBridge,
+    XChainModifyBridgeFlag,
+    XChainModifyBridgeFlagInterface,
+)
 
 __all__ = [
     "AccountDelete",
     "AccountSet",
     "AccountSetFlag",
     "AccountSetFlagInterface",
+    "AMMBid",
+    "AMMCreate",
+    "AMMDeposit",
+    "AMMDepositFlag",
+    "AMMDepositFlagInterface",
+    "AMMVote",
+    "AMMWithdraw",
+    "AMMWithdrawFlag",
+    "AMMWithdrawFlagInterface",
+    "AuthAccount",
     "CheckCancel",
     "CheckCash",
     "CheckCreate",
@@ -98,4 +139,14 @@ __all__ = [
     "TrustSet",
     "TrustSetFlag",
     "TrustSetFlagInterface",
+    "XChainAccountCreateCommit",
+    "XChainAddAccountCreateAttestation",
+    "XChainAddClaimAttestation",
+    "XChainClaim",
+    "XChainCommit",
+    "XChainCreateBridge",
+    "XChainCreateClaimID",
+    "XChainModifyBridge",
+    "XChainModifyBridgeFlag",
+    "XChainModifyBridgeFlagInterface",
 ]